heredoc refactored
