+++ Makefile
@@ -9,7 +9,10 @@ C_DIR = src
 
 C_FILES = debug.c \
 	main.c \
-	free.c \
+	free/minishell.c \
+	free/envvar.c \
+	free/token.c \
+	free/cmd.c \
 	exit.c \
 	error.c \
 	env/init.c \
@@ -22,21 +25,23 @@ C_FILES = debug.c \
 	tokenizer/tokenizer_utils.c \
 	tokenizer/token_utils.c \
 	tokenizer/quote_handler.c \
-	parser/init_cmd_list.c \
+	parser/cmd/init_list.c \
+	parser/cmd/new_cmd.c \
+	parser/cmd/utils.c \
 	parser/parser.c \
 	parser/parser_utils.c \
-	parser/cmd_utils.c \
 	parser/handlers/handle_word.c \
 	parser/handlers/handle_pipe.c \
 	parser/handlers/handle_redir_in.c \
 	parser/handlers/handle_redir_out.c \
 	parser/handlers/handle_redir_append.c \
 	parser/handlers/handle_redir_heredoc.c \
-	parser/heredoc.c \
 	execute/parent_process.c \
 	execute/child_process.c \
 	execute/pipe.c \
-	execute/redirection.c \
+	execute/redirection/save_restore.c \
+	execute/redirection/setup.c \
+	execute/heredoc.c \
 	execute/executable.c \
 	execute/builtin.c \
 	execute/builtins/cd.c \
