+++ inc/minishell.h
@@ -59,20 +59,28 @@ typedef struct s_token
 	struct s_token	*next;
 }	t_token;
 
+typedef struct	s_heredoc
+{
+    char	*delimiter;
+    int		fd;
+}	t_heredoc;
+
 typedef struct s_cmd
 {
 	char			**args;
-	char			*infile;
-	char			*outfile;
-	char			*heredoc_del;
-	int				heredoc_fd;
-	int				append;
+
+	pid_t			pid;
 	int				*pipe;
 	int				saved_stdin;
-	int				fdin;
 	int				saved_stdout;
+
+	int				fdin;
 	int				fdout;
-	pid_t			pid;
+	char			*infile;
+	char			*outfile;
+	t_heredoc		*heredoc;
+	int				append;
+
 	struct s_cmd	*prev;
 	struct s_cmd	*next;
 }	t_cmd;
@@ -116,10 +124,11 @@ void			put_error(char *fmt, ...);
 
 /* Memory */
 void			free_minishell(t_minishell **minishell);
-void			free_token_lst(t_token **token_lst);
-t_envvar		*free_envvar_node(t_envvar **node);
 void			free_envvar_lst(t_envvar **var_lst);
+t_envvar		*free_envvar_node(t_envvar **node);
+void			free_token_lst(t_token **token_lst);
 void			free_cmd_lst(t_cmd **cmd_lst);
+t_cmd			*free_cmd_node(t_cmd **cmd);
 
 /* Exit */
 void			exit_minishell(int exit_code, t_minishell *minishell,
@@ -146,6 +155,7 @@ void			put_signal_message(int status);
 void			kill_all_children(t_minishell *ms);
 int				get_and_reset_signal(void);
 t_bool			ft_signal(int signum, void (*handler)(int));
+int	get_heredoc_signal(void); //DEBUG
 
 /* Parsing */
 void			init_cmd_lst(t_minishell *minishell);
@@ -153,9 +163,7 @@ t_cmd			*cmd_new(t_cmd *prev_cmd);
 void			add_arg_to_cmd(t_cmd *cmd, char *arg);
 int				parse_tokens(t_minishell *minishell);
 t_parse_op		*get_parse_ops(void);
-int				process_heredoc(t_cmd *cmd, t_minishell *ms);
-//int			process_all_heredocs(t_minishell *ms); //TODO
-//void			cleanup_heredoc(t_minishell *ms); // TODO keep?
+int				process_all_heredocs(t_minishell *ms);
 void			handle_token_type(t_token **cur_token, t_cmd **cur_cmd,
 					t_minishell *minishell);
 void			handle_redir_in(t_token **cur_token, t_cmd **cur_cmd,
