--- src/execute/heredoc.c.bak
@@ -1,180 +1,0 @@
-#include "minishell.h"
-
-/**
- * Generates a unique temporary file for heredoc input.
- *
- * Creates a file with the pattern `/tmp/minishell_heredoc_<counter>` where:
- * - `<counter>` is a static integer incremented for each heredoc
- * - File is created with exclusive access (O_EXCL Requires 
- *   that the file must not exist before opening)
- * - File permissions are 0600 (user read/write) for security 
- *
- * @param tmp_file Buffer where the generated filename is stored.
- * (minimum 256 bytes)
- * @return File descriptor of the created file, or -1 on failure.
- */
-static int	generate_unique_file(char *tmp_file)
-{
-	static int	counter;
-	const char	*base;
-	char		*num_str;
-	int			fd;
-
-	while (1)
-	{
-		num_str = ft_itoa(counter++); // TODO free itoa
-		if (!num_str)
-			return (1);
-		base = "/tmp/minishell_heredoc_";
-		ft_strlcpy(tmp_file, base, 256); // TODO handle failure + free
-		ft_strlcat(tmp_file, num_str, 256); // TODO handle failure + free
-		ft_free(1, &num_str);
-		fd = open(tmp_file, O_WRONLY | O_CREAT | O_EXCL, 0600); // TODO handlefailure
-		if (fd != -1)
-			return (fd);
-	}
-}
-
-/**
- * Reads user input for heredoc (<<) and writes it to the specified file.
- * 
- * Continuously reads input from the user until the heredoc delimiter 
- * (cmd->heredoc_del) is encountered. Each input line is written to the 
- * temporary file. If the delimiter is encountered, input stops.
- * 
- * @param tmp_fd File descriptor for the temporary heredoc file.
- * @param cmd Pointer to the command struct containing the heredoc delimiter.
- * @return EXIT_SUCCESS after successfully reading the heredoc input.
- */
-static int	read_heredoc(int tmp_fd, t_cmd *cmd)
-{
-	char	*line;
-
-	while (1)
-	{
-		line = readline("> ");
-		if (!line)
-		{
-			ft_free(1, &line); // TODO useful?
-			return (EXIT_FAILURE);
-		}
-		if (ft_strcmp(line, cmd->heredoc_del) == 0)
-		{
-			ft_free(1, &line);
-			break ;
-		}
-		ft_fprintf(tmp_fd, "%s\n", line);
-		ft_free(1, &line);
-	}
-	return (EXIT_SUCCESS);
-}
-
-/**
- * Processes the heredoc (<<) redirection by storing input to a tmp file.
- *
- * - Calls @generate_unique_file to create a temporary file
- * - Calls @read_heredoc in order to read input until delimiter is entered
- * - Stores file path in command structure
- *
- * @param cmd Pointer to the command struct containing heredoc details.
- * @return EXIT_SUCCESS on success, EXIT_FAILURE if file creation or memory 
- * allocation fails.
- */
-static int	process_heredoc_in_child(t_cmd *cmd)
-{
-	char	tmp_file[256];
-	int		tmp_fd;
-
-	tmp_fd = generate_unique_file(tmp_file);
-	if (tmp_fd == -1)
-		return (EXIT_FAILURE);
-	if (read_heredoc(tmp_fd, cmd) != EXIT_SUCCESS)
-		return (EXIT_FAILURE);
-	ft_close(&tmp_fd);
-	cmd->infile = ft_strdup(tmp_file);
-	cmd->heredoc_tmpfile = ft_strdup(tmp_file);
-	if (!cmd->infile || !cmd->heredoc_tmpfile)
-	{
-		unlink(tmp_file);
-		ft_free(2, &cmd->infile, &cmd->heredoc_tmpfile);
-		return (EXIT_FAILURE);
-	}
-	return (EXIT_SUCCESS);
-}
-
-int	process_heredoc(t_cmd *cmd)
-{
-	int	status;
-
-	pid_t pid = fork();
-	if (pid < 0) {
-		put_error("fork");
-		return (EXIT_FAILURE); // TODO
-	}
-	if (pid == 0) {
-		signal(SIGINT, SIG_DFL);
-		exit(process_heredoc_in_child(cmd)); //TODO return (process_heredoc_in_child(cmd));
-	}
-	else
-	{
-		waitpid(pid, &status, 0);
-		if (WIFEXITED(status))
-			return (WEXITSTATUS(status));
-		if (WIFSIGNALED(status))
-			return (E_SIGBASE + WTERMSIG(status));
-	}
-	return (EXIT_SUCCESS);
-}
-
-/**
- * Processes heredoc (`<<`) redirections for all commands in the command list.
- *
- * Iterates through each command in `ms->cmd_lst`, checking if it has a 
- * heredoc delimiter (`heredoc_del`). If so, it calls process_heredoc.
- *
- * If any heredoc processing fails, `ms->exit_code` is set to EXIT_FAILURE.
- *
- * @param ms Pointer to the global Minishell structure.
- * @return EXIT_SUCCESS (failure is tracked in `ms->exit_code`).
- */
-int	process_all_heredocs(t_minishell *ms)
-{
-	t_cmd	*cmd;
-	int		exit_code;
-
-	cmd = ms->cmd_lst;
-	while (cmd)
-	{
-		if (cmd->heredoc_del)
-		{
-			exit_code = process_heredoc(cmd);
-			if (exit_code != EXIT_SUCCESS)
-			{
-				ms->exit_code = exit_code;
-				return (EXIT_FAILURE);
-			}
-		}
-		cmd = cmd->next;
-	}
-	return (EXIT_SUCCESS);
-}
-
-/*
- * Deletes heredoc_tmpfile and cleans up the relevant resources.
- */
-void	cleanup_heredoc(t_minishell *ms)
-{
-	t_cmd	*cmd;
-
-	cmd = ms->cmd_lst;
-	while (cmd)
-	{
-		if (cmd->heredoc_tmpfile)
-		{
-			unlink(cmd->heredoc_tmpfile);
-			ft_free(1, &cmd->heredoc_tmpfile);
-			cmd->heredoc_tmpfile = NULL;
-		}
-		cmd = cmd->next;
-	}
-}
