+++ src/parser/cmd/init_list.c
