+++ src/utils/string.c
